Remove the karma junit reporter
We may have used it in our jenkins tests at some point but we don't
have those anymore. It weas pulling in ancient dependencies because
we were using version 2.0.0 which is fact much older than the
current version (1.2.0). We have little use for junit output anymore
so just remove it.