Fix regression on widget panel edit button
Due to commit https://github.com/matrix-org/matrix-react-sdk/commit/6895fb72d41cd1349a1e81585266d86a6fa168c3
the parameters for the call to open a widget in edit mode from the
widget panel in a room has broken. The `screen` parameter needs
to be prefixed with `type_` as it was before.

This corresponds to parameters supplied when creating the URL when
opening a widget in edit mode through Scalar screens.

Signed-off-by: Jason Robinson <jasonr@matrix.org>