Guard against falsy names in getInitialLetter
This ensures we check for a falsy name in `getInitialLetter` instead of throwing
errors. We should perhaps also fix whatever other issues have led to the input
being undefined in the first place, but for now we leave this for another day.

Hopefully helps with https://github.com/vector-im/riot-web/issues/10983