Fix the build
Run `yarn install` in the js-sdkc and react-sdk installs. We were
avoiding this before due to using babel dependencies from riot-web
ro build js-sdk, which worked as long as the installed babel versions
and plugins were identical. js-sdk has now been upgraded to babel 7
and so that no longer worked.

This will probably cause react-sdk & js-sdk to be built an
unnecessary number of times.